version 4.0

+++ models/ProcessPost.php
@@ -13,7 +13,17 @@
 
 		public function getAllPost(){
 			$sql = "SELECT * FROM posts";
-			return $this->select($sql);
+			$result = $this->select($sql);
+			foreach ($result as $post) {
+				$post->content = $this->makeReadmore($post->content,$post->id,$post->url);
+			}
+			return $result;
+		}
+
+		public function getFivePost(){
+			$sql = "SELECT title,id,url FROM posts LIMIT 5";
+			$result = $this->select($sql);
+			return $result;
 		}
 
 		public function getPost($post_id){
@@ -33,18 +43,32 @@
 			WHERE id='$post_id' AND userid='$userid' ";
 			return $this->update($sql);
 		}
-		public function deletePost($id){
-			$sql = "DELETE FROM posts WHERE id='$id'";
+		public function deletePost($id,$grantUser,$userID){
+			if($grantUser == "0"){
+				$sql = "DELETE FROM posts WHERE id='$id'";
+			}else{
+				$sql = "DELETE FROM posts WHERE id='$id' AND userid='$userID'";
+			}
 			return $this->delete($sql);
 		}
 		public function getPostofCategory($category_id){
 			$sql = "SELECT * FROM posts WHERE category_id = '$category_id'";
-			return $this->select($sql);
+			$result = $this->select($sql);
+			foreach ($result as $post) {
+				$post->content = $this->makeReadmore($post->content,$post->id,$post->url);
+			}
+			return $result;
+		}
+
+		public function deleteCategory($id){
+			$sql = "DELETE FROM categories WHERE id='$id'";
+			return $this->delete($sql);
 		}
 
 		public function getCategory($category_id){
 			$sql = "SELECT * FROM categories WHERE id = '$category_id'";
-			return $this->select($sql);
+			$result = $this->select($sql);
+			return $result[0];
 		}
 
 		public function getCategoryId($category_link){
@@ -63,6 +87,11 @@
 			$sql = "INSERT INTO categories(title,url,description) VALUES ('$title','$url','$description')";
 			return $this->insert($sql);
 
+		}
+
+		public function updateCategory($title,$url,$description,$id){
+			$sql = "UPDATE categories SET title='$title', url = '$url',description='$description' WHERE id = '$id'";
+			return $this->update($sql);
 		}
 	}
